+++ src/di/tests/Stub/Inject/Foo3Trait.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://hyperf.wiki
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ */
+namespace HyperfTest\Di\Stub\Inject;
+
+use Hyperf\Di\Annotation\Inject;
+
+trait Foo3Trait
+{
+    use FooTrait;
+
+    /**
+     * @Inject
+     * @var Bar
+     */
+    protected $bar;
+
+    public function getValue()
+    {
+        return 'foo3';
+    }
+}