+++ src/rpc-client/src/ServiceClient.php
@@ -58,6 +58,10 @@ class ServiceClient extends AbstractServiceClient
         $response = $this->checkRequestIdAndTryAgain($response, $id);
         if (array_key_exists('result', $response)) {
             $type = $this->methodDefinitionCollector->getReturnType($this->serviceInterface, $method);
+            if ($type->allowsNull() && $response['result'] === null) {
+                return null;
+            }
+
             return $this->normalizer->denormalize($response['result'], $type->getName());
         }